Abstract getPluginsConfig
